CausalMixGPD: Dirichlet Process Mixtures with Generalized Pareto Tail

CausalMixGPD provides Bayesian Dirichlet process mixture models for unconditional, conditional, and causal workflows, with optional generalized Pareto tail modeling.

Details

Core entry points include build_nimble_bundle(), run_mcmc_bundle_manual(), build_causal_bundle(), and run_mcmc_causal().

Learn more via:

package index: help(package = "CausalMixGPD")

vignettes: browseVignettes("CausalMixGPD")

website articles index: https://arnabaich96.github.io/CausalMixGPD/pkgdown/articles/index.html

vignette pages:

vignette("01-basic", package = "CausalMixGPD")

vignette("03-unconditional", package = "CausalMixGPD")

vignette("04-conditional", package = "CausalMixGPD")

vignette("05-causal", package = "CausalMixGPD")

website: https://arnabaich96.github.io/CausalMixGPD/

See also

Author

Maintainer: Arnab Aich aaich@fsu.edu